refactor ff tenery and rename util file
