+++ step-generation/src/utils/pythonFileUtils.ts
